refactor(automation): split discovery and status out of the auto runner
runner.py had grown to 997 lines against a 1000-line hard limit, holding
three unrelated jobs: finding candidates, reporting durable status, and
executing one attempt under a lease.

The first two move out along their own boundaries:

- candidate_discovery.py — CandidateDiscovery turns one live library
  snapshot into ranked candidates (detect, back off, apply variety, score).
  It takes Config plus two Protocol read seams, MemoryHistoryReader and
  FailureStreakReader, by constructor injection.
- status.py — the read-only status contract (AutomationStatus,
  CooldownStatus, SuggestStatus) and the cooldown gate, which reads
  through a CompletedRunReader seam.

runner.py keeps what only it can do: the OS lease, the generation
subprocess and its bounded redacted output, and the attempt record.

Pure move, no behavior change. The runner still publishes its last_*
state in the same order, so a failed library read leaves the recent
categories intact. The recent-auto-runs query that status and suggest
each ran separately is now one _recent_auto_runs helper with the same
arguments.

Line counts: runner.py 997 -> 603, candidate_discovery.py 328,
status.py 192.

The complexipy snapshot drops AutoRunner::suggest (was 21, over the 15
limit) because the split retired it — the ratchet tightening that the
Makefile target describes.
